Move ProjectileProperties
