+++ api-internal/src/main/java/at/pavlov/internal/projectile/ProjectileProperties.java
@@ -1,4 +1,4 @@
-package at.pavlov.cannons.projectile;
+package at.pavlov.internal.projectile;
 
 public enum ProjectileProperties
 {
